backend/requirements.txt
celery==5.6.3
dj-database-url==3.1.2
django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.17.1
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.3.0
openpyxl==3.1.5
pandas==3.0.1
pillow==12.1.1
psycopg[binary]==3.2.12
pymupdf==1.27.2.2
pytesseract==0.3.13
python-dotenv==1.2.2
rapidfuzz==3.14.3
redis==7.4.0
requests==2.33.0
